Reorganize to use upstream tarball
Given this is a packaging of software from an upstream, and not
software written entirely for Debian/Ubuntu, it should use non-native
packaging.

Reference: https://www.debian.org/doc/manuals/maint-guide/first.en.html#non-native-dh-make

The goal of this commit is to build a deb with exactly the same
contents as the original PPA deb, but using the filesystem layout
of the upstream tarball. This also drops the unnecessary toplevel
makefile and replaces it with debhelper config.

Verified accuracy using:

$ dpkg -c old.deb
    sed -r 's/(.*) [0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2} (.*)/\2 \1/' | \
    sort > old.list
$ dpkg -c new.deb
    sed -r 's/(.*) [0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2} (.*)/\2 \1/' | \
    sort > new.list
$ diff -u old.list new.list
$