Generate a policy-confirming copyright file
The current copyright file does not point the reader to where the
original source was obtained as required by policy. It also refers
to an external document for copyright and licensing information,
while policy requires this text be included verbatim.

Reference: https://www.debian.org/doc/debian-policy/ch-docs.html#s-copyrightfile

Fix this by generating a copyright file at build time that contains:
 a) A static header which points users to the download source
 b) The contents of the SPDX file, which appears to contain the
    necessary copyright and licensing information.

The original copyright file pointed to a .htm file as a reference,
but the SPDX format seems to include the same information and is
more suitable for command line usage.