Remove "Replaces" relationship on self
There is no need to declare a "Replaces" relationship on earlier versions of
the same package. Installation of a new version automatically results in the
removeal of older versions. Coinstallation of multiple versions of the same
package/arch is not supported.

Reference: https://www.debian.org/doc/debian-policy/ch-relationships.html#s-replaces